fix: primaryask追加発行に対応
